some love for Fredi, some hash to anchors
I figured that if editing somewhere in scrolled page it wouldn't jump/stay to the location after refreshing page. This worked in Firefox but not in Chrome for some reason. I added a hash to links to then add to window.location.

Not sure if the issue in Chrome is a different, but maybe addition like this could make sense?